making codebase more intuitive stable
