--- hemutwebapp/app/globals.css
@@ -1,56 +1,0 @@
-
-.app {
-  display: flex;
-}
-
-.main-content {
-margin-left: 240px;
-width: 100%;
-min-height: 100vh;
-background-color: #f5f5f5;
-transition: margin-left 0.3s ease, width 0.3s ease;
-}
-
-.main-content.expanded {
-margin-left: 0px;
-width: 100%;
-}
-
-.page-content {
-font-family: 'Poppins', sans-serif;
-width: 100%;
-margin: 0 auto;
-min-height: 100vh;
-display: flex;
-flex-direction: column;
-align-items: center;
-gap:4px;
-}
-
-html, body {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  height: 100%;
-  width: 100%;
-}
-
-*, *::before, *::after {
-  box-sizing: inherit;
-}
-
-#root, #__next {
-  height: 100%;
-  width: 100%;
-}
-
-
-/* scrolling */
-body {
-  /* overflow: hidden; */
-  font-family: 'Poppins', sans-serif;
-}
-
-/* .page-content::-webkit-scrollbar {
-  display: none;
-} */
