--- hemutwebapp/app/logout/layout.tsx
@@ -1,19 +1,0 @@
-import "../globals.css";
-export const metadata = {
-  title: "Logout",
-  description: "Logout Screen",
-};
-
-export default function LogoutLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body>
-        <main>{children}</main>
-      </body>
-    </html>
-  );
-}
