+++ hemutwebapp/app/logout/page.tsx
@@ -1,4 +1,5 @@
 "use client";
+// import "../styles/globals.css";
 import "../styles/loginpage.css";
 import HemutLogo from "../assets/HemutLogo";
 import Card from "../components/login-components/Card";
@@ -28,7 +29,7 @@ const LogoutPage = () => {
           <AccountPrompt
             message="Don't have an account?"
             linkText="Sign up"
-            href="/logout/signup"
+            href="/signup"
           />
         </LogIn>
       </div>
