+++ hemutwebapp/app/styles/globals.css
@@ -1,0 +1,28 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap');
+html, body {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  height: 100%;
+  width: 100%;
+}
+
+*, *::before, *::after {
+  box-sizing: inherit;
+}
+
+#root, #__next {
+  height: 100%;
+  width: 100%;
+}
+
+
+/* scrolling */
+body {
+  /* overflow: hidden; */
+  font-family: 'Poppins', sans-serif;
+}
+
+/* .page-content::-webkit-scrollbar {
+  display: none;
+} */
