+++ hemutwebapp/app/styles/home.css
@@ -14,3 +14,13 @@ transition: margin-left 0.3s ease, width 0.3s ease;
 margin-left: 0px;
 width: 100%;
 }
+.page-content {
+  font-family: 'Poppins', sans-serif;
+  width: 100%;
+  margin: 0 auto;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap:4px;
+  }
